+++ pytest_unique/unique.py
@@ -1,4 +1,5 @@
 """Unique data generation."""
+
 import string
 import sys
 import time
@@ -8,7 +9,7 @@ from pathlib import PurePath
 from random import sample
 from uuid import UUID
 
-from attrs import define, field
+from attr import define, field
 
 from pytest_unique.count import memory_count
 from pytest_unique.registry import registry_load
@@ -116,7 +117,7 @@ def unique_password(unique, lowercase=4, uppercase=2, digits=1, punctuation=1):
     return "".join(sample(chars, lowercase + uppercase + digits + punctuation))
 
 
-def unique_text(unique, prefix=None, suffix=None, separator="-"):
+def unique_text(unique, prefix=None, suffix=None, separator="-", limit=None):
     """Return text unique to this factory instance.
 
     :param prefix: Used as a prefix for the unique string, defaults
@@ -124,6 +125,7 @@ def unique_text(unique, prefix=None, suffix=None, separator="-"):
     :param suffix: Optional suffix for the unique string.
     :param separator: Separator between parts of the string,
         defaults to '-'.
+    :param limit: Optional limit for the unique string.
     """
     if prefix is None:
         frame = sys._getframe(2)
@@ -136,18 +138,20 @@ def unique_text(unique, prefix=None, suffix=None, separator="-"):
             .replace("_", separator)
             .replace(".", separator)
         )
-        prefix = separator.join(
-            [
-                "unique",
-                "from",
-                source,
-                f"line{frame.f_lineno}",
-            ]
-        )
+        prefix = separator.join([
+            "unique",
+            "from",
+            source,
+            f"line{frame.f_lineno}",
+        ])
 
     text = separator.join([prefix, str(unique("integer"))])
     if suffix is not None:
         text = separator.join([text, suffix])
+
+    if limit is not None:
+        # Truncate from the end.
+        text = text[-limit:]
 
     return text